Move Zustand store to FormManger level; will also split into slices, as appropriate.
